requirement.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
psycopg2-binary
bcrypt
